+++ src/components/InputsUserInfo/InputProfileImage/InputProfileImage.tsx
@@ -3,8 +3,8 @@ import React, { useCallback, useState } from 'react';
 import { useDropzone } from 'react-dropzone';
 import defaultProfileImage from '../../../assets/images/defaultProfile.png';
 import { Theme } from '../../../styles/Theme';
-import Button from '../../Button/Button';
-import Label from '../../Label/Label';
+import Button from '../../common/Button/Button';
+import Label from '../../common/Label/Label';
 import ProfileImage from '../../Profile/ProfileImage/ProfileImage';
 import {
   profileImageSelectContainer,
