+++ src/components/InputsUserInfo/InputPw.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { css } from '@emotion/react';
-import Input from '../Input/Input';
+import Input from '../common/Input/Input';
 import { ERROR } from '../../constants/message';
 
 const InputPw = () => {
