+++ src/components/common/Button/Button.style.ts
@@ -1,6 +1,6 @@
 import { css } from '@emotion/react';
 
-import { Theme } from '../../styles/Theme';
+import { Theme } from '../../../styles/Theme';
 
 import type { ButtonProps } from './Button';
