+++ src/components/common/Input/Input.style.ts
@@ -2,7 +2,7 @@ import { css } from '@emotion/react';
 
 import type { InputProps } from './Input';
 
-import { Theme } from '../../styles/Theme';
+import { Theme } from '../../../styles/Theme';
 
 export const inputContainerStyling = css({
   display: 'flex',
