+++ src/components/common/Label/Label.style.ts
@@ -1,6 +1,6 @@
 import { css } from '@emotion/react';
 
-import { Theme } from '../../styles/Theme';
+import { Theme } from '../../../styles/Theme';
 
 export const labelStyling = css({
   fontSize: Theme.text.small.fontSize,
